fortnite-booot/requirements.txt
discord.py==2.3.2
Flask